Issue #541
Fixed missing tooltips.

The issue was around raising the tolltip.

In process, merged class code into main appjar code, and removed tooltip from library.
